+++ puffin_http/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to `puffin_http` will be documented in this file.
 
 
 ## Unreleased
+* Better handle slow clients, especially when there are multiple clients.
+
+
+## 0.5.1 - 2021-09-16
+* Fix high-bandwidth connection interruptions.
 
 
 ## 0.5.0 - 2021-09-06